+++ run_tests.sh
@@ -925,6 +925,7 @@ ptest_program "examples/switch_string_return_test.om" 7
 ptest_program "examples/struct_static_call_test.om" 11
 ptest_program "examples/top_const_test.om" 10
 ptest_program "examples/paren_free_test.om" 12
+ptest_program "examples/round71_test.om" 0
 flush_ptests
 
 echo ""